+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-Cxqpl1Iv.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-CM_TjcZT.js";import{_ as a}from"./mermaid.core-Bmo0IU8w.js";import"./chunk-5VM5RSS4-CAJG7VIV.js";import"./chunk-XXDRQBXY-nOZDaTTj.js";import"./chunk-POPQ4Y6H-BlhYv_Hz.js";import"./chunk-F27PBJKO-dEeCUlf0.js";import"./channel-CioVe4MR.js";import"./index-BzaxTmp-.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};